Wrap expectEqual to fix type inference
The builtin `std.testing.expectEqual(expected, actual)` breaks type
inference for optionals and integers by forcing the `actual` value
to match the type of the `expected` value instead of vice-versa:
this causes code like `expectEqual(2, int_variable)` to fail to compile,
insisting that `int_variable` must be a `comptime_int` rather than
interpreting `2` as whatever type `int_variable` has.

Workarounds are to explicitly cast the expected value with
`@as(actual_type, expected_value)`, which is clunky, or to flip the
order of `actual` and `expected`, which causes misleading errors
on failed assertions.

This wrapper fixes this while preserving the intended parameter order.

See https://github.com/ziglang/zig/issues/4437#issuecomment-683309291
for further discussion.