Adopt Type/Instance structural style for Program and VirtualMachine
For source files that declare struct types:
- The source file should be imported as `TypeName` (e.g. const Program = @import("program.zig")`.
- The struct type inside it is referred to as `TypeName.Instance` (e.g. `Program.Instance`), and never imported directly.
- Any namespace-level constants are referred to as `Type.constant_name` or `Type.Subtype` (e.g. `Program.Error`).
- Functions which generate new instances should live at the top level of the namespace, instead of nested inside the instance type: e.g. `Program.init` instead of `Program.Instance.init`.
- Instance methods should remain namespaced within the Instance type: e.g. `Program.Instance.read`. This makes it clear at a glance what is an instance method and what is a type-level method.

This style came out of my dissatisfaction with zig's `lowercase_with_underscores` style for namespaces, which very easily collided with local variable names and led to awkward constructions like `program.Program.init`. I plan to roll it out to the rest of the files in subsequent commits.

(Since zig files are really structs, an alternative style is to eliminate the nested `Instance` type altogether and declare member variables at the top level of the file; I tried this but found it made it very unclear what the member variables of the struct actually were, and references within the file to its own type needed to be aliased with const `Self = @This();` which was a little too idiomatic for me.)