Adopt new naming style for Instruction wrapped type
This semantic change is important because the Wrapped union is not the primary way machine instructions should be consumed; rather, the virtual machine will eventually parse-and-execute bytecode in a single step to avoid an intermediate conversion step, which is wasteful in space (unions are as large as their largest member) and time (wrapping and then executing an instruction would be a half-assed form of dynamic dispatch).

The Wrapped type is primarily intended decompiling and analyzing Another World bytecode resources.