Extract ControlThreads operation type to its own source file
I'm of two minds about this, because the type is only relevant to one specific operation.

However, it has enough logic and its own tests; and my approach to defining parse/new functions at the top level of a file does not play nice with having multiple "hero" types in the same source file.