Implement resource descriptors and MEMLIST.BIN parser
The calling conventions for `parse`, which reads a list of descriptors out of a MEMLIST.BIN file, are subject to change. Currently the caller is expected to provide a buffer of sufficient size to store the expected number of descriptors; in future the `parse` function may take an ArrayList instead, or take an allocator and return a buffer for the caller to own.