Implement RLE instruction parsing
I made a separate parser that consumes a bitwise reader and returns RLE instructions for a writer to consume. This made it easier to test than implementing the parser as part of `run_length_decoder.Reader` itself; constructing CRC-correct reversed fixture data is a nightmare I do not wish to undertake. We'll cover real-world data parsing in integration tests once the writer is implemented.

We're introducing some inefficiency for the sake of testability: `readInstruction` returns a union type that encapsulates each possible instruction to be consumed later by a writer, whereas an efficient implementation would talk directly to the writer without the intermediate translation.

This will not practically matter on a modern CPU, especially since we're only decompressing assets at scene load.