Add checksum status for reader
This is an attempt to reflect that the CRC checksum of the reader is only relevant once it has finished reading.

We could also represent this with a two-case union, where one case is `data_remaining` and the other is `finished` with a boolean. But in zig, that takes more one more byte than a tri-state enum and is more cumbersome to consume.