Implement writer for run-length decoder
The logic for relative offsets ended up a little squirrely on account of the write cursor being 1 higher than the element it will actually write to.

This was done to ensure that a cursor of 0 means "the buffer is full", as it does for the reader; since cursor is unsigned, it cannot go to -1 without an underflow, and so we'd have to set some separate flag / use some sentinel value to mark "we're done here".

Since the cursor should not be consumed outside of the Writer type, I think that's fine.