Implement run-length decoder
This puts all the pieces together: it combines the reader, parser and writer into a single function, which is called from the resource loader when it encounters compressed data.

All of these components are unit-tested individually, but are hard to test all together with fake data; so I'm relying on the integration tests for now to parse real-world resources without errors. Unfortunately this will not catch data corruption errors, such as the writer writing the bytes out of order.