Refactor RLE decoding to get rid of Parser wrapper
Previously we had a parser type that wrapped a bitwise reader and parsed it for instructions, which it reified into an intermediate union type that a top-level decode function had to interpret and pass along to the writer.

This refactor gets rid of the parser and the union type, in favor of a function that takes a reader and writer and calls the appropriate methods on both directly. The benefits:
- We remove the intermediate indirection of creating a union type, which was a waste of time and bytes.
- We avoid the awkwardness of having a type (Parser) that depended on two other mutable types but only wrapped one of them.
- We've moved all of the instruction-handling logic from the top-level decode function into more unit-testable places, so all the real work is happening on a single layer of abstraction.

To preserve the testability we used to get from the Instruction union, there is now a mock writer type that listens for the appropriate methods and generates a union from them.

We still have the top-level decode function, which probably needs to live somewhere else; but it now knows very little about the implementation, instead stitching together the real (not mock) types and making them talk nice to each other. Rather than inject mocks, I'll cover this function with a couple of fixture tests in a future PR.

We also still ended up needing a wrapper type for `readInt` and `readByte` functions, which parse a sequence of bits from a bitwise reader. This is stateless behaviour that we wanted to share between multiple reader types (the real thing and the mock); in Swift we'd implement those in a protocol extension, but zig has no equivalent means to share code other than global namespace functions (which don't play nice with duck typing, making it hard to substitute mock types). I'm still not happy with this and will try to find some more elegant pattern.