Make RLE reader validate its own completion state
This moves more logic out of the top-level decode function, and eliminates the unnecessary Status enum which needed to be converted into an error anyway. (Validation functions that return `!void` feel like a useful pattern in zig for this kind of thing.)

It also brings the mock reader's and wrapper reader's API in line with the real bit reader.