Make real and mock RLE readers wrap themselves in an interface
Since the underlying readers could not be used without wrapping them first, and since ideally callers should be unaware of the wrapping, it makes sense for those types to wrap themselves.

ReaderInterface is effectively an interface type, with no other than the reader it wraps. Interfaces do not have a solid standard pattern in Zig yet; this is one possibility.

One benefit of this refactor is that it became harder to access properties of the underlying reader, forcing the declaration of a clear public interface. This was a good thing; it meant that tests had to test behaviour (isAtEnd()) rather than implementation (cursor == 0).