Clarify RLE decoding error behaviour
Previously, `SourceBufferNotFullyConsumed` had a misleading name and could only ever occur if the uncompressed buffer size was undercounted by just the right amount that decoding otherwise appeared to complete successfully.

By enforcing instead that both the reader and writer must be exhausted, decoding behaviour becomes much easier to reason about, and this edge case becomes clearly a mistake in usage rather than a result of corrupt data. As a result we can give it and its validator function a clearer name.