Add function for executing arbitrary instructions from a program
This is intended to be used in the VM's main loop to bypass the parse-to-union step.

Since all switch branches have the same calling conventions, we could create a lookup array of function pointers instead like the original C reference implementation did; but this switch statement is likely more efficient (no pointer dereference needed) and not any wordier.