Add placeholder interfaces for audio and resource-loading subsystems
This eliminates logging statements from the instructions in favor of calls to the machine itself: this removes dependencies on stdlib and brings the instruction implementations close to their final form.

Importing functions into the namespace of another struct is a pattern I haven't seen before in Zig code. I'm not sure how I feel about it; it will help to break up larger implementations, but you can no longer see the full interface of a struct at a glance.

This arrangement is temporary in any case, and will be reevaluated once I start to implement these subsystems for real. Ideally audio, video, resources etc. would be their own structs nested within a larger machine; but it is very likely that these subsystems will each need to access and modify state of the virtual machine as a whole, which gets messy with nesting. To be continued.