Add DrawBackgroundPolygon instruction
This and the other (yet-to-be-implemented) polygon instruction are special cases, in that they do not map to a single opcode byte value, but are identified with the top two bits of the opcode byte and use the remaining bits of the opcode as part of the instruction. This required some more squirrely parsing logic.

Fortunately this logic could be implemented entirely in the Opcode type, isolating the rest of the instruction code from having to care about it. Go science!