Use mock machine to test DrawSpritePolygon.Instance.execute
Mocks are necessary for testing machine functions whose side-effects are hard to measure otherwise, such as producing audio or rendering pixels on a screen.

I rather like this approach, which builds a mock instance out of a struct that implements the functions that are expected to be called. The mock provides compile-time proof of which functions the function under test is calling; if you fail to implement one of them in the Stubs struct, it will result in a compile error.