Add tests for MockMachine
This demonstrates a possible pattern for counting function invocations. Inner functions cannot access local variables of the surrounding function, so we must define the invocation count as a global within a named type in order to access its value.

We might be able to tweak this pattern such that fields in the implementation struct could be surfaced directly on the mock instance; I haven't found a good way yet, but `usingnamespace` has potential.