Make metaprogramming compatible with latest zig syntax
Somewhere in the last 4 weeks, Zig's builtins were overhauled to remove all type introspection fields from `type` in favor of the builtin `TypeInfo` API.

This made a lot of the code to e.g. get the bit width of an integer or the return type of a function maddeningly verbose, so I've wrapped those in helpers.