Import reader methods directly onto reader types instead of wrapping
This is a similar pattern to Ruby modules, where a type imports declarations from another type into its own namespace.

I'm of two minds about this pattern - I don't like the way Ruby modules hide where functionality is defined. But, avoiding wrapping removes the need to define passthrough methods for each public method on the wrapped type.

On the other hand, it also means there's no single concrete type that functions can be constrained to, which is one advantage of the Reader and Writer interface types in the stdlib.