Add DrawString instruction
This is a straight port of the reference implementation and I'm not super-happy with how it does an O(n) search on the string table. I'd like to check how many gaps there are in the table; it may be quicker, tidier and more compact just to define a contiguous array and leave nulls for the gaps.