Minor cleanup
I'd like to avoid initializing structs as undefined as much as possible, but there's several that are still initialized piecemeal because the logic for each field is complex and failable.

I should look into what instructions the compiler will generate if you assign those to local variables and construct the struct at the end; it may be that it's smart enough to reuse the return struct's destination fields as storage for the local variables that will eventually be "copied" into those fields.