Remove optional scale in favor of default constant
This removes an unnecessary type complexity and the ambiguity of `64` and `null` having the same behaviour.

I kept the `default` union case in DrawSpritePolygon for now because I don't *think* the extra case takes any more space and it makes the fixture definitions a little nicer. It could easily be replaced with `.{ .constant = Video.default_scale }` though.