Create library build target and reorganize project structure
This is the result of running `zig init-lib` and moving everything into src like it seems to want. The main.zig file currently does nothing but reference as many files as are necessary to reach all the tests.

This is a step on the way to having Github's CI build and test for me.