Selectively disable zig fmt for some code blocks
This allows me to commit the results of `zig fmt` without selectively undoing all the time.

I gave up and let `zig fmt` undo my fancy indentation in a lot of places, where column-aligning code had not made it any easier to understand the code - e.g. because the code needed to be read row-wise to be understood anyway.