Add RegisterSubtract instruction
Editor's note: we could consolidate the implementations of all these register instructions into two concrete struct types - one that has two registers, and one that has a register and a constant - and use an operation enum (a la JumpConditional) to decide what to do with the values. This would unify the tests and cut down on a bunch of boilerplate.

This is something to consider once we have naive implementations of all of these instructions; I don't want to prematurely optimize.