Add model for bounding boxes
These will be used for fast is-polygon-on-screen checks when rendering.

I'm not super-happy with the truncation and wrapping going on here. Another World tracked screen coordinates as signed 16-bit values, and polygon scales could theoretically boost coordinates or bounding box dimensions out of range and cause them to overflow. But a bounding box that overflows will have unexpected behaviour, and I don't think the original game ever relied on coordinate wrapping.

One solution is to stick to 16-bit coordinates and treat overflows as a runtime error case; another is to widen all our internal coordinates to `isize` and prevent wrapping altogether (since a 16-bit coordinate multiplied by a 16-bit scale cannot overflow 32 bits).