Preliminary implementation of video buffer
The VideoBuffer class will be responsible for implementing polygon and glyph rendering, deferring to an underlying storage type for the actual reading and writing of pixels.

The abstraction allows different storage implementations to be compared. The original game packed two 16-bit pixels into a single byte, which saved a lot of space but complicated draw routines by requiring a lot of masking to read and write those pixels. A buffer that uses a single byte per pixel will be simpler to implement and test, and may (or may not!) be faster, at the cost of twice the space; I want to performance-test the two approaches for myself.