Preliminary implementation of span drawing
This is a component of the polygon rendering algorithm, which renders polygons row by row as horizontal spans between interpolated vertices.

Drawing a span pixel-by-pixel at the VideoBuffer layer is less efficient than it would be at the storage layer, where fixed-color drawing could be implemented as a memset and masked drawing could be implemented as a buffer-to-buffer copy. I'll explore that approach later.