Add indexed bitmap type for testing video buffers
`expectEqual` and `expectEqualSlices` are not great for testing multidimensional arrays; defining the expected value is tedious, and the error messages they give back are cryptic and unhelpful.

This Bitmap type represents the contents of a 2D video buffer in a way that can be easily converted to/from a multiline string representation; this lets us define expected values more concisely, and take advantage of the far better error messages produced by expectEqualStrings.