Use packed structs instead of shift+truncate
This makes it a little cleaner to read individual bits out of integers. One wrinkle is that fields in a packed struct are ordered from least to most significant bits. I’m not sure whether this is affected by big vs little-endian; though that should only affect the order of bytes and not the order of bits within a byte.