Fix test failure in resource decoder
At some point I had refactored the decoder to preflight-check that the destination has enough bytes left to unpack the source. This changed the error produced by a test that accidentally declared a write operation larger than the destination - what it really wanted to test was a write operation larger than the source.