Remove single-pixel draw functions for aligned storage
These had a bug that the unit tests missed, where the single-pixel highlight draw function tried to access a (nonexistent) mask buffer. I could have fixed this, but the fact that this got missed felt like an argument for getting rid of the special case; it increased complexity and reduced testability for a micro-optimization whose branching would probably just slow things down in practice.