Preliminary implementation of video subsystem
This contains the video state needed (palette, buffers etc) and draw routines that depend on data beyond a specific buffer. It lacks tests and the following functionality is still missing:
- String drawing
- Copying from buffer to buffer
- Rendering to host screen

The idea is for Machine.Instance to eventually contain and initialize its own Video.Instance struct, setting its polygons and animations to appropriate pointers into a memory block managed by the machine itself. This separation of concerns is practical because none of the draw routines in Video.Instance depend on data in registers or threads.

Still undecided is whether bytecode instructions like DrawBackgroundPolygon should talk directly to Video.Instance rather than through the Machine interface. The problem is with mocking it in tests - it could greatly complicate the MockMachine to also mock a nested structure with its own methods. (Potentially compile-time reflection could cut out a lot of the existing boilerplate in MockMachine, which would make it more viable.)