Prevent trapping in bounding box calculations
Previously the use of @intCast would trap when the dimensions were too large, allowing pathological cases to reach undefined behaviour and crash the app. Changing this to use native-width integers fixes this, but it also reveals a pre-existing pathological case with overflows that it will handle differently from the original game.

Since the game probably never expected to deal with values large enough to overflow anyway, I can live with undefined rendering behaviour instead of a runtime crash.