Treat most polygon errors as non-fatal when parsing
Previously we did full validation on every polygon that was parsed from game data. This was unnecessary work, because the draw algorithm is actually robust against many polygon errors that represent bad data and would fail validation. Instead we now do the bare minimum of validation during parsing and use validation only in the integration tests.