Add parser for Another World palette resources
This sparked an exploration into the unusual colorspace mapping used by the reference implementation, and the likely historical cause for it. For now I've decided to match the reference implementation, but I suspect it would be more correct - and much more simple to just multiply the values out.

I also learned that the palette resources (at least in the DOS version) were twice as large as expected, because they stored EGA as well as VGA palettes. For now we only parse the VGA palettes; it will be interesting later to parse both, and allow the emulator to toggle between the two graphics modes.