Refactor resource loader to support loading into existing buffer
This is in aid of keeping a fixed memory layout for the emulator with no dynamic allocation: instead we'd define static buffers of the same size as the original game used for loading resource data, and replace their contents when moving from one game part to another.

One place we may still need dynamic allocation is file-path handling, since modern OSes can have paths of arbitrary length. For this reason I've left the resource path handling alone for now.