Fix ControlThreads panicking when iterating full range of threads
This occurred because ControlThreads was implicitly using the ThreadID.Trusted type (which is u6) as its counter. After the loop had reached thread 63 (the maximum value for ThreadID.Trusted), it tried to increment again and overflowed.

This was only discovered by attempting to execute all bytecode instructions in the game in an integration test. There are probably similar bugs where we loop over a value with a tightly-constrained type.