Remove explicit error type from instruction wrapper
This was a lot of boilerplate for very little gain. Zig was able to resolve the full set of possible errors at compile time, and error sets this large are not useful as documentation. Making the error type implicit reduces the amount of boilerplate we need to maintain in that hideous file.