Add helper for fixed-capacity buffers with dynamic length
This abstracts the pattern used by polygons into a reusable type. I'm not sure how I feel about this and fear it may be an anti-pattern. It's effectively a copyable slice that carries around its own storage instead of using an allocator. I would expect to see something like this in the standard library if it was a blessed way of doing things in manually memory-managed languages.

An alternative would be storing a slice alongside the buffer it points to, but that would not be safely copyable or relocatable (in the same way as any struct that contains a pointer to itself).