Eliminate memory allocation in Resource Loader path handling
Instead of holding onto an allocator to build absolute paths out of filenames and to parse the resource list, the resource loader now uses:
- fixed buffers for generating DOS filenames, which have a guaranteed maximum length;
- directory handles for opening files;
- a fixed-capacity resource list array.

This doesn't really improve the ergonomics of the class: you still need to deinit it to close its directory handle, and references to the loader now need to be `var` so that the directory handle can unset itself on close. But this does feel like a cleaner and more correct way of using the filesystem APIs.