Avoid allocation when resolving fixture game directory
By making the Resource Loader no longer "own" its directory handle, acquiring the handle could be moved into the test helpers and we can avoid doing any string manipulation to get absolute paths.

This improves the ergonomics of the tests - one fewer `defer` and the loader can be const - but means that the resource loader can no longer enforce that the directory handle's lifetime will match its own. RAII is not a strong pattern in Zig, but this feels like a step backward in safety.

One approach would be ResourceLoader/ResourceLoaderUnmanaged, a la ArrayList, but that would add a bunch of boilerplate to let the former encapsulate the latter. I shall revisit this decision once resource loading is wired up within the emulator itself and we can judge the ergonomics there.