Rename ResourceLoader to ResourceDirectory
This is to more consistently name types in the codebase after *what concept they represent*, rather than *what task they perform*.

Also, I wanted the freedom to reuse the "ResourceLoader" name for a new type that would read data from disk using a ResourceDirectory and keep track of where the data loaded in memory. That would more closely fits the notion of a "loader".