Load palettes dynamically instead of all at once
Palettes are now kept in their raw format in memory and converted into 24-bit host palettes on the fly, instead of converting them all at once.

This matches the reference implementation and  the way other resources are handled, allows conversion to other color formats, and saves memory; but it's also slower and significantly complicates the code. I will reevaluate this decision later once rendering is working end-to-end.