Rename Repository interface to Reader
The previous naming convention introduced ambiguity between the repository interface and the concrete implementation of that interface, resulting in us sometimes calling the implementations "data sources" instead of "repositories".

Renaming the interface to "Reader" follows the Zig standard library interfaces on files and byte streams (which have the same conventions around ownership and lifetime) and makes the relationship between repository and reader clearer in code.

The downside is that it may cause confusion with the actual Zig I/O Reader class. If this becomes a problem, we can come up with another name for the interface type (e.g. "Loader") and continue to reserve the "Repository" term for the long-lived parent types.