Rename startGamePart to scheduleGamePart
The reference implementation defers starting a new game part until the end of the current frame so that all scripts have time to finish, so we need to separate the scheduling of the new game part from the actual loading of it. Scheduling is not failable, so the instruction that fires it no longer needs to do error handling.