Preliminary implementation of virtual machine game part loading
This is a very partial implementation, enough to change the external API but not the internals. It is still missing tests and I'm not particularly happy with the resource-passing API between the machine and the video subsystem.

To prevent the virtual machine from ever being in an illegal partially-initialized state, the machine attempts to load a game part as soon as it is created and discards the VM instance if that fails. This means that every place that needs a machine also needs to provide a valid repository.

That affected the tests for a couple of dozen instruction types, so I've added a test machine helper for unit tests that don't want to mock the machine and don't care about proper resource loading.