Add virtual machine host interface
This interface will eventually get methods for audio output; for now though, it concerns itself just with frame rendering.

In the current design, the VM requests a suitably-sized buffer from the host and writes 24bpp color data into it, rather than passing a palette and a 4bpp buffer to the host and letting it do the conversion. This is less flexible and denies the host certain opportunities for optimization (e.g. paletted hardware textures); but it insulates the host from the internal details of our buffer format, which is still in flux. We can revisit this design easily later.