Replace var with const for const references to mutable data
Const pointers to mutable data still allow that data to be mutated; the pointer only needs to be `var` if the pointer itself can be modified to point to different memory. One day, hopefully the zig compiler or linter will catch vars-that-could-be-const for me.

I hadn't just marked pointers as var unnecessarily, but I/O readers too. It took me a while to grasp how a construction like `const reader = fixedBufferStream(ptr).reader();` is still legal. `io.Reader` does not have any mutable state of its own, so can be `const`, but its parent buffer stream holds a read/write cursor that must be mutated when the reader is read. It seems that within a function, Zig detects that the `reader()` method requires a mutable callee, and implicitly marks the callee's region of the stack as mutable, equivalent to the following code:
```zig
var stream = fixedBufferStream(ptr);
const reader = stream.reader();
```

However, that doesn't seem to apply for global variables. e.g. the following does *not* compile when it is at global scope:
```zig
var buffer: [16]u8 = undefined;
const global_reader = std.io.fixedBufferStream(&buffer).reader();
```
In this situation, zig treats the parent buffer stream as const unless you explicitly assign it to a static `var`.