Parse palette immediately on selection
This avoids the need to parse the palette anew every single frame.

It also introduces a new edge case: trying to render to a surface before a palette has been selected (i.e. before the current game part's bytecode has executed a selectPalette instruction. We could avoid this edge case by always loading palette 0 whenever the palette resource is changed, but that may be redundant and the reference implementation didn't default to any specific palette.

We could also avoid this edge case by setting the palette to `undefined` and just rendering garbage if we ever reach that state; but I'd rather find out first if the real game ever tries to render in that state.

To satisfy that curiosity, we currently treat attempting to render before there's a palette as an error.