Add explicit signed/unsigned API for registers, and convert RegisterID type into an enum
Previously we treated register values as signed by default and bit-cast them to unsigned when needed. This was expressive but it left the common case ambiguous: it was unclear from a casual reading whether a given context *knew* it was working with a signed value and was handling it appropriately.

Wrapping the registers in a type with explicit getters and setters makes it obvious at every call site what the caller intends; while it prevents modification-in-place expressions like `registers[id] += 1`, it shouldn't actually result in more cycles than the former (and we can always use a pointer-based API if we really mind this).

The change to explicit getters also made it possible to represent register IDs as a non-exhaustive enum instead of a size-constrained integer. This has a number of advantages:
- Debugging: printing a RegisterID will indicate the name of a known enum instead of the raw value, making it easier to see what a program is doing.
- Correctness: Register IDs cannot be accidentally passed to a parameter expecting a regular integer or vice-versa, e.g. if someone messes up the order of arguments to a function.
- Correctness: likewise, you cannot perform arithmetic on register IDs.

(A non-exhaustive enum in Zig is analogous to a "distinct type" in languages like Nim, see e.g. https://nim-by-example.github.io/types/distinct/ for further discussion.)

However, these two changes combined do reduce the ergonomics of accessing register values - especially in tests. I may change my mind and revert all this.