Replace separate input->register methods with registerValues struct
The previous method-per-register implementation was conceptually elegant but did a lot of unnecessary work, repeatedly testing the same conditions.

By combining them into a single method, we do all the work once; and the unit tests serve as a good visual model for how the various registers will change in sync with each other.