Clarify intent of last_character_typed
My original implementation was based on a misunderstanding of the reference implementation. It was zeroing out the keycode in its input struct after consuming it, which meant that it would set the register to zero on the very next input frame.

To accomplish this in our VM, we expect the host to `null` out the `last_character_typed` field if no key was released during the current input loop, and should 0 out the register in response to `null`.