Make MockRepository generate valid bytecode programs
This works filling the resource buffer with the simplest possible legal program: yield instructions followed by a jump back to start. This can be safely executed by the VM for an arbitrary number of tics without hitting an error condition.

There may be merit in abstracting this from MockRepository into a "fixture generator". However, there would probably be more merit in redesigning MockRepository's API to allow you to provide a pointer to arbitrary data for a given descriptor or resource type.