Add benchmark for intro cinematic
This is very rough-and-ready at the moment, and the analysis is copypasta from https://github.com/ziglang/gotta-go-fast/blob/master/bench.zig. I still understand very little about the Builder API, so the new benchmark build target may be set up in a clunky way.

So far the median time to run through the intro at max speed is ~0.25sec in Debug mode, ~0.13sec in Release-Safe mode and ~0.09sec(!) in Release-Fast mode.

Future improvements to the benchmark will be to run as many iterations as possible in [n] seconds instead of running [n] iterations; and to abstract the measurements into a type that can be reused for other benchmarkable things (e.g. rendering every polygon in the game).