Render frames to surface in benchmark test
This produces a more meaningful test of the VM.
With this, we see that:
Debug build takes ~700ms to complete;
Release-safe Takes ~97ms;
Release-fast takes ~91ms.

I discovered I had miscounted the significant digits in the previous benchmark implementation: without the render-to-surface stage, release-safe took only 13ms to complete the demo, and release-fast only 9ms. It's possible that the compiler was aggressively optimizing away the results of draw operations after noticing the results were never read.

Implementing reveals that the original game attempts to render before a palette has been selected; either the game happily rendered garbage for a handful of frames, or it defaulted the palette to e.g. index 0 or all black, and I (or the reference implementation) had just missed that. Further investigation needed.