Fix thread state leaking between game parts
Previously we were not resetting the scheduled state of threads when a new game part was loaded - only the current thread state.

This caused program counter corruption when switching from the password entry screen to a new game part: the password entry screen would schedule a jump on a thread on the tic before the part transition, which activated that thread on the first tic of the new game part and jumped to an offset that was invalid in the new game part's program. This caused invalid instruction errors at best, graphical corruption and program hangs at worst.