Add SDL2 frontend
The VM now actually runs, using SDL2 to render to a 640x480 window and receive keyboard input.

To try this out on MacOS, run the following commands in the project directory:
1. Run `brew install sdl2` to install the SDL2 library via Homebrew. (https://brew.sh)
2. Ensure you have put MS-DOS Another World game files in fixtures/dos.
3. Run `zig build run` to launch the game.

(This should work on Linux by skipping step 1, as long as Zig can otherwise locate the SDL2 library. You may need to use the `--search-prefix` argument to tell `zig build` where to find it.)

The VM currently crashes when starting the next game part after the introduction: it will terminate with an out of bounds array access error owing to an off-by-one mistake with bounding box calculations.