Fix crash when drawing intersecting
This was afflicting the first screen of the first gameplay part of the game, which was trying to draw a single-dot polygon whose bounds just intersected, but wasn't enclosed by, the screen bounds. This occurred because single-dot polygons have a bounding box 2 rows high even though only the top row is drawn, so when only the bottom row was technically in bounds the top row would be drawn out of bounds and crash.

The fix was to always consider single-dot polygons as a point rather than a bounding rect. Later I may reify this distinction into the Polygon.Instance type itself as a union, so that a single-dot polygon doesn't even have a bounding box.